create new in order index removed, new product a price can be chosen and in the index is shown the finalPrice

+++ maison_belmain_symfony/src/Entity/ProductPrice.php
@@ -14,19 +14,19 @@ class ProductPrice
     private ?int $id = null;
 
     #[ORM\Column]
-    private ?int $amount = null;
+    private ?float $amount = null;
 
     public function getId(): ?int
     {
         return $this->id;
     }
 
-    public function getAmount(): ?int
+    public function getAmount(): ?float
     {
         return $this->amount;
     }
 
-    public function setAmount(int $amount): static
+    public function setAmount(float $amount): static
     {
         $this->amount = $amount;
